1-1-bzhosko : create html

+++ hotel-app/src/main/java/by/mycom/ita/services/impl/CommonUserServiceImpl.java
@@ -35,7 +35,7 @@ public class CommonUserServiceImpl implements ICommonUserService {
 
     @Override
     public CommonUser findById(Long id) {
-        return userDao.findById(id).orElseThrow(DataNotFoundException::new);
+        return userDao.findById((long)id).orElseThrow(DataNotFoundException::new);
     }
 
 }
